refactor: split the task row codec and the store contract tests at their seams
`coerceTask` and its per-field helpers move to `task-coerce.ts`; `store-codec.ts`
keeps the manifest-level codec (versions, tombstones, disk merge). The
store-load-edge test keeps load() recovery and the runtime contract cases move
to `store-contract-edge.test.ts`. Both files had crossed the 500-line cap.
